% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_table.R
\name{object_table}
\alias{object_table}
\title{Create an object table}
\usage{
object_table(get = NULL, set = NULL, has = NULL, unbind = NULL,
  names = NULL, attach = NULL, detach = NULL, parent_env = emptyenv())
}
\arguments{
\item{get}{`function(name) {}`. Can return any type of object. Affects
the operation of `$`, `[[`, `get()`.}

\item{set}{`function(name, value) {}`. Should return logical indicating
success or failure. Affects the operation of `$<-`, `[[<-` and `assign`.}

\item{has}{`function(name) {}`. Should return logical. Affects the
operation of [exists()].}

\item{unbind}{`function(name) {}`. Should return logical indicating success
or failure. Affects the operation of [rm()].}

\item{names}{`function() {}`. Should return character vector.}

\item{attach, detach}{`function() {}`. Called when environment is attached
or detached. Can be used to run set up and tear down code}

\item{parent_env}{This environment is used to implement default operations
for all other arguents.}
}
\description{
An object table is a custom environment that allows you to redefine the
usual behaviour of R's environments.
}
\examples{
ot <- object_table(get = function(name) nchar(name))
ot$a
ot$this_is_a_long_name
}
